Update and rename Default.java to Svenska.java
Swedish translation by SnickarN.
